feat: debounce page builder saves and centralize plugin types
