+++ packages/plugins/paypal/package.json
@@ -7,6 +7,7 @@
     ".": "./index.ts"
   },
   "dependencies": {
+    "@acme/types": "workspace:*",
     "zod": "^3.25.73"
   }
 }
