+++ packages/types/src/index.ts
@@ -21,3 +21,4 @@ export * from "./Coverage";
 export * from "./upgrade";
 export * from "./ExampleProps";
 export * from "./Segment";
+export * from "./plugins";
